+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # World In a Jar 1.0.0-beta.4+1.21.1.blanketcon
+## Added
+- Jars can now be locked. They are locked by default and can be unlocked using the Glowcase Lock item.
+
 ## Fixed
 - Only send a single block update to clients surrounding a jar when a BlockState is set.